Python/DjangoTests/fusion/requirements.txt
asgiref==3.2.3
coverage==5.0.4
dj-static==0.0.6
Django==3.0.4
django-stdimage==5.0.2
gunicorn==20.0.4
model-mommy==2.0.0
Pillow==7.0.0
progressbar2==3.50.0
psycopg2-binary==2.8.4
python-utils==2.4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
